tests/requirements.txt
pytest
web3
py-solc-x
eth-tester
